Better status condition management (fix findings) #41
