+++ src/main/java/io/strimzi/kafka/access/internal/StatusUtils.java
@@ -2,7 +2,7 @@
  * Copyright Strimzi authors.
  * License: Apache License 2.0 (see the file LICENSE or http://apache.org/licenses/LICENSE-2.0.html).
  */
-package io.strimzi.kafka.access.model;
+package io.strimzi.kafka.access.internal;
 
 import java.time.ZoneOffset;
 import java.time.ZonedDateTime;
